Fix bd entities

+++ BESA/AgenteResPwA/ResPwAEntities/src/ResPwAEntities/Joint.java
@@ -23,7 +23,7 @@ import javax.xml.bind.annotation.XmlTransient;
 
 /**
  *
- * @author juans
+ * @author maria.f.garces.cala
  */
 @Entity
 @Table(name = "JOINT")
@@ -50,7 +50,7 @@ public class Joint implements Serializable {
     private double angulo;
     @Basic(optional = false)
     @Column(name = "TIEMPO")
-    private BigDecimal tiempo;
+    private BigInteger tiempo;
     @ManyToMany(mappedBy = "jointList", fetch = FetchType.EAGER)
     private List<Accion> accionList;
 
@@ -61,7 +61,7 @@ public class Joint implements Serializable {
         this.id = id;
     }
 
-    public Joint(BigDecimal id, String nombre, double angulo, BigDecimal tiempo) {
+    public Joint(BigDecimal id, String nombre, double angulo, BigInteger tiempo) {
         this.id = id;
         this.nombre = nombre;
         this.angulo = angulo;
@@ -92,11 +92,11 @@ public class Joint implements Serializable {
         this.angulo = angulo;
     }
 
-    public BigDecimal getTiempo() {
+    public BigInteger getTiempo() {
         return tiempo;
     }
 
-    public void setTiempo(BigDecimal tiempo) {
+    public void setTiempo(BigInteger tiempo) {
         this.tiempo = tiempo;
     }
